+++ __init__.py
@@ -2,4 +2,4 @@ __author__ = "Ian Housman"
 __email__ = "ian.housman@gmail.com"
 
 # Version format yyyy.m.n
-__version__ = "2024.5.3"
+__version__ = "2024.8.3"