+++ docs/source/conf.py
@@ -1,7 +1,7 @@
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here.
 import pathlib
-import sys, os, re, shutil
+import sys, os, re, shutil, glob
 
 # sys.path.insert(0, os.path.abspath("..Examples/"))
 # sys.path.insert(0, os.path.abspath(".."))
@@ -25,6 +25,40 @@ def GetVersion(initPath):
 
 build_folder = r"A:\GEE\gee_py_modules_package\geeViz\docs\build"
 info_folder = r"A:\GEE\gee_py_modules_package\geeViz\docs\source\info"
+
+# Manage notebooks
+notebook_folder_input = r"A:\GEE\gee_py_modules_package\geeViz\examples"
+notebook_folder_output = r"A:\GEE\gee_py_modules_package\geeViz\docs\source\notebooks"
+examples_template = r"A:\GEE\gee_py_modules_package\geeViz\docs\source\templates\examples_template.rst"
+examples_out = r"A:\GEE\gee_py_modules_package\geeViz\docs\source\examples.rst"
+sync_notebooks = True
+update_examples_template = False
+notebooks = glob.glob(os.path.join(notebook_folder_input, "*.ipynb"))
+
+# Update notebooks to be included if specified
+if sync_notebooks:
+    if os.path.exists(notebook_folder_output):
+        shutil.rmtree(notebook_folder_output)
+        os.makedirs(notebook_folder_output)
+
+    for nb in notebooks:
+        out_nb = os.path.join(notebook_folder_output, os.path.basename(nb))
+        print(out_nb)
+        shutil.copy(nb, out_nb)
+if update_examples_template:
+    notebooks = glob.glob(os.path.join(notebook_folder_output, "*.ipynb"))
+    notebooks = [os.path.splitext(os.path.basename(n))[0] for n in notebooks]
+    notebook_string = "\n\n\t".join("notebooks/" + n for n in notebooks)
+    o = open(examples_template, "r")
+    example_lines = o.read()
+    o.close()
+    o = open(examples_out, "w")
+    example_lines = example_lines.replace("{NOTEBOOK_PATHS}", notebook_string)
+
+    o.write(example_lines)
+    o.close()
+
+# Clean up existing build
 if os.path.exists(build_folder):
     shutil.rmtree(build_folder)
 if os.path.exists(info_folder):
@@ -35,32 +69,26 @@ project = "geeViz"
 copyright = "2024, Ian Housman, Josh Heyer"
 author = "Ian Housman, Josh Heyer"
 release = GetVersion(initPath)
-
+nb_execution_mode = "off"
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
-extensions = [
-    "sphinx.ext.autodoc",
-    "sphinx.ext.autosummary",
-    "sphinx.ext.coverage",
-    "sphinx.ext.napoleon",
-]
-
-templates_path = ["_templates"]
-exclude_patterns = []
+extensions = ["sphinx.ext.autodoc", "sphinx.ext.autosummary", "sphinx.ext.coverage", "sphinx.ext.napoleon", "sphinx.ext.viewcode", "myst_nb", "sphinx_copybutton"]
 
 
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
-
+print("here", os.path.dirname(os.path.realpath(__file__)))
 html_theme = "furo"
 html_static_path = ["_static"]
-
+html_css_files = ["style/custom.css"]
+# html_style = "style/custom.css"
 dark_text_color = "#00bfa5"
 light_text_color = "#00897b"
 
-html_logo = f"./{html_static_path[0]}/RCR-logo.jpg"
-html_favicon = html_logo
+html_logo = "images/RCR-logo.jpg"
+html_favicon = "favicon.ico"
+# html_favicon = html_logo
 pygments_style = "sphinx"
 pygments_dark_style = "monokai"
 
@@ -88,7 +116,7 @@ html_theme_options = {
         {
             "name": "RedCastle Resources Inc",
             "url": "https://www.redcastleresources.com/",
-            "html": f"<img src='{html_logo}'  alt='RedCastle Inc. Logo' href='#' title='Click to learn more about RedCastle Resources Inc.'>",
+            "html": f"<img src='https://apps.fs.usda.gov/lcms-viewer/src/assets/images/RCR-logo.jpg'  alt='RedCastle Inc. Logo' href='#' title='Click to learn more about RedCastle Resources Inc.'>",
             "class": "",
         },
     ],